standalone boot
git-svn-id: http://192.168.42.160/software/main@17566 ce24552a-085e-4656-8849-d4bfd34db3c6

+++ board/mv_ebu/common/USP/mv_ide.c
@@ -1371,7 +1371,7 @@ ulong dma_read_write_cmd (MV_SATA_ADAPTER *pSataAdapter,
 
 	if ( blkcnt > 128){
 		printf("error in %s: blk count %llx exceeded max limit\n",
-		       __func__, blkcnt);
+		       __func__, (long long unsigned int)blkcnt);
 	}
 
 	//	mvSataDisableChannelDma(pSataAdapter, channelIndex);
